backend/pyproject.toml
[project]
name = "audienceroom-backend"
version = "0.1.0"
description = "AudienceRoom backend API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "python-dotenv>=1.0.1",
  "pydantic[email]>=2.7.0",
  "firebase-admin>=6.5.0",
  "google-generativeai>=0.8.0",
  "openai>=1.50.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]